# docs/sdd/openapi.yaml
openapi: 3.0.3
info:
  title: Mango Trading API
  description: Simulador de trading com ações fictícias — contrato REST desacoplado. Baseado em controller/* e readme.md.
  version: 0.0.1
  contact:
    name: Mango Trading
servers:
  - url: http://localhost:8080
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: "Ana Silva" }
        email: { type: string, format: email, example: "ana@email.com" }
        password: { type: string, minLength: 6, example: "123456" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    TradeRequest:
      type: object
      required: [stockId, quantity]
      properties:
        stockId: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1, example: 10 }
    StockRequest:
      type: object
      required: [name, symbol, currentPrice]
      properties:
        name: { type: string }
        symbol: { type: string, example: "MANG3", description: "Uppercased, unique, 3-10 alphanum" }
        description: { type: string }
        sector: { type: string, example: "Tecnologia" }
        currentPrice: { type: number, format: double, example: 42.50, minimum: 0.01 }
        volatility: { type: number, format: double, example: 0.025, minimum: 0 }
        active: { type: boolean, example: true }
    MarketEventRequest:
      type: object
      required: [title, impact]
      properties:
        title: { type: string }
        description: { type: string }
        impact: { type: string, enum: [LOW, MEDIUM, HIGH] }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        recurrenceType: { type: string, enum: [NONE, DAILY, RANDOM], example: "RANDOM" }
        dailyTime: { type: string, example: "09:00", description: "HH:mm for DAILY" }
        randomMinMinutes: { type: integer, example: 30 }
        randomMaxMinutes: { type: integer, example: 120 }
        randomMin: { type: string, example: "30m", description: "alternative e.g. 30m, 2h" }
        randomMax: { type: string, example: "2h" }
        durationMinutes: { type: integer, example: 60 }
        scope: { type: string, enum: [ALL, SECTOR], example: "ALL" }
        sector: { type: string, example: "Tecnologia" }
        priceDeltaPercent: { type: number, format: double, example: 3.5 }
        direction: { type: string, enum: [POSITIVE, NEGATIVE, RANDOM], example: "RANDOM" }
        enabled: { type: boolean, example: true }
        visibilityMinutes: { type: integer, example: 3 }
        daysOfWeek: { type: string, example: "MONDAY,TUESDAY" }
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string }
        balance: { type: number, format: double, example: 10000.00 }
        role: { type: string, enum: [USER, ADMIN] }
        createdAt: { type: string, format: date-time }
    JwtResponse:
      type: object
      properties:
        token: { type: string }
        type: { type: string, example: "Bearer" }
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string }
        role: { type: string }
    StockResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: "Mango Tech" }
        symbol: { type: string, example: "MANG3" }
        description: { type: string }
        sector: { type: string }
        currentPrice: { type: number, format: double }
        volatility: { type: number, format: double }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MarketEventResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string }
        impact: { type: string, enum: [LOW, MEDIUM, HIGH] }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        recurrenceType: { type: string, enum: [NONE, DAILY, RANDOM] }
        dailyTime: { type: string, example: "09:00" }
        randomMinMinutes: { type: integer }
        randomMaxMinutes: { type: integer }
        durationMinutes: { type: integer }
        scope: { type: string, enum: [ALL, SECTOR] }
        sector: { type: string }
        priceDeltaPercent: { type: number, format: double }
        direction: { type: string, enum: [POSITIVE, NEGATIVE, RANDOM] }
        enabled: { type: boolean }
        nextRunAt: { type: string, format: date-time }
        lastRunAt: { type: string, format: date-time }
        visibilityMinutes: { type: integer }
        daysOfWeek: { type: string }
        status: { type: string, enum: [SCHEDULED, UPCOMING, ACTIVE, ENDED] }
        secondsToStart: { type: integer }
    TransactionResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        stockId: { type: integer, format: int64 }
        stockSymbol: { type: string }
        stockName: { type: string }
        type: { type: string, enum: [BUY, SELL] }
        quantity: { type: integer }
        price: { type: number, format: double }
        createdAt: { type: string, format: date-time }
    PortfolioItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        stockId: { type: integer, format: int64 }
        stockSymbol: { type: string }
        stockName: { type: string }
        sector: { type: string }
        currentPrice: { type: number, format: double }
        quantity: { type: integer }
        averagePrice: { type: number, format: double }
        totalInvested: { type: number, format: double }
        currentValue: { type: number, format: double }
        profitLoss: { type: number, format: double }
        profitLossPercent: { type: number, format: double }
    PortfolioResponse:
      type: object
      properties:
        portfolioId: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        userName: { type: string }
        balance: { type: number, format: double }
        items:
          type: array
          items: { $ref: '#/components/schemas/PortfolioItemResponse' }
        totalInvested: { type: number, format: double }
        totalCurrentValue: { type: number, format: double }
        totalPatrimony: { type: number, format: double }
        totalProfitLoss: { type: number, format: double }
    RankingResponse:
      type: object
      properties:
        position: { type: integer }
        userId: { type: integer, format: int64 }
        userName: { type: string }
        balance: { type: number, format: double }
        stockValue: { type: number, format: double }
        totalPatrimony: { type: number, format: double }
    ErrorEnvelope:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        message: { type: string }
        errors:
          type: object
          additionalProperties: { type: string }
security: []
paths:
  /api/auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '400':
          description: Validation or duplicate email
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JwtResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/stocks:
    get:
      summary: List active stocks
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/StockResponse' } }
        '401': { description: Unauthorized }
  /api/stocks/all:
    get:
      summary: List all stocks
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/StockResponse' } }
        '401': { description: Unauthorized }
  /api/stocks/{id}:
    get:
      summary: Stock detail
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockResponse' }
        '404': { description: Not found }
        '401': { description: Unauthorized }
  /api/events:
    get:
      summary: List market events (UPCOMING/ACTIVE, 3min visibility)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MarketEventResponse' } }
        '401': { description: Unauthorized }
  /api/events/all:
    get:
      summary: List all market events (admin view)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MarketEventResponse' } }
        '401': { description: Unauthorized }
  /api/trades/buy:
    post:
      summary: Buy stock
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TradeRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '400': { description: Insufficient balance or inactive }
        '404': { description: Stock not found }
        '401': { description: Unauthorized }
  /api/trades/sell:
    post:
      summary: Sell stock
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TradeRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '400': { description: Insufficient quantity }
        '404': { description: Not owned }
        '401': { description: Unauthorized }
  /api/trades/history:
    get:
      summary: Trade history
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TransactionResponse' } }
        '401': { description: Unauthorized }
  /api/portfolio:
    get:
      summary: Portfolio
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioResponse' }
        '404': { description: Not found }
        '401': { description: Unauthorized }
  /api/ranking:
    get:
      summary: Ranking
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/RankingResponse' } }
        '401': { description: Unauthorized }
  /api/admin/stocks:
    get:
      summary: Admin list stocks
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/StockResponse' } }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
    post:
      summary: Admin create stock
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StockRequest' }
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockResponse' }
        '400': { description: Duplicate symbol }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
  /api/admin/stocks/{id}:
    get:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        '404': { description: Not found }
    put:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/StockRequest' } } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockResponse'
        '404': { description: Not found }
    delete:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '204': { description: No content }
        '404': { description: Not found }
  /api/admin/events:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarketEventResponse'
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MarketEventRequest' } } }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketEventResponse'
        '400': { description: Bad request }
  /api/admin/events/{id}:
    get:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketEventResponse'
        '404': { description: Not found }
    put:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MarketEventRequest' } } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketEventResponse'
    delete:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '204': { description: No content }
  /api/admin/events/{id}/preview:
    get:
      summary: Preview next runs
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: count, in: query, required: false, schema: { type: integer, default: 3 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { type: string, format: date-time } }
        '404': { description: Not found }
  /api/admin/events/{id}/trigger:
    post:
      summary: Trigger event now
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarketEventResponse' }
        '404': { description: Not found }
  /api/admin/users:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
  /api/admin/users/{id}:
    get:
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404': { description: Not found }